+++ src/server/settings/settings-router.ts
@@ -1,8 +1,9 @@
 import { router } from "../trpc/trpc"
-import { get, save } from "./settings-procedures"
+import { checkForUpdate, get, save } from "./settings-procedures"
 
 export let settings = router({
     settings: {
+        checkForUpdate,
         get,
         save
     }